Moved Modulator Functions to one Common File
