--- bin/modulator/master.js
@@ -1,6 +1,0 @@
-module.exports = {
-    load: require('./load'),
-    launch: require('./launch'),
-    install: require('./install'),
-    update: require('./update')
-}
